pr changes

+++ src/services/nandi/nandi_router.py
@@ -150,7 +150,7 @@ def validate_rate_feedback_api(
     rate_id: str,
     feedbacks: list[str] = [],
     chargeable_weight: int = None,
-    price: int = None,
+    price: float = None,
     currency: str = None,
     resp: dict = Depends(authorize_token)
     ):
